refactor(db): move credential/custom_provider repos into djinn-db
`credential.rs` (24 raw sqlx calls) and `custom_provider.rs` (4) were the
last two files violating scripts/check-raw-sql-boundary.sh. They passed CI
only because the guard is diff-scoped and neither file was being touched.

This is the completion of a half-finished refactor, not a design change.
ef02b5104 moved crypto + these two repositories out of djinn-db and into
djinn-provider. 4c78658c4, the SAME DAY, moved the crypto half back "so
CredentialRepository can access crypto without depending on the provider
crate". The repositories never followed. No ADR defends the placement, and
`dispatch_pause.rs` — the identical shape — has always lived in djinn-db.

Nothing blocked the move: djinn-provider -> djinn-db is the only edge, and
the two files imported nothing from djinn-provider outside one test.

Also drops djinn-provider's sqlx dependency entirely. It declared
`features = ["mysql", ...]` while every query in repos/ is Postgres; the
crate could never have built its own queries on its own declared feature
set, and only compiled because workspace-hack unifies features workspace-
wide. sqlx appeared nowhere else in the crate, so the dependency is gone
rather than corrected. (The bogus "mysql" feature also appears in
djinn-agent, djinn-core, djinn-memory and the workspace root; those are
left alone here.)

Call sites do not churn: djinn_provider::repos is now a re-export shim, so
all ~37 referencing files are untouched.

One test could not move. `user_model_priority_keeps_openai_via_codex_merge`
needs djinn_provider::catalog::CatalogService, which djinn-db must not
depend on. It moved the other way, into
djinn-provider/tests/credential_user_model_priority.rs, where both halves
are in scope. Test count is unchanged: 18 before (15 + 3), 18 after
(14 + 3 + 1).

The guard is untouched — no exemption was added. It now passes honestly.

Verification:
- Guard fails on origin/main for both files (exit 1, 24 + 4 hits);
  whole-tree run on this branch is clean (1259 files checked, exit 0).
- `.sqlx` did not churn: `git status server/.sqlx` empty, and
  `make sqlx-verify` reports "up to date (408 entries)".
- `cargo check --workspace --all-targets --all-features` clean.
- `cargo clippy -p djinn-db -p djinn-provider --all-targets --all-features
  -- -D warnings` clean; `cargo fmt --all --check` clean.
- `cargo hakari verify` passes (workspace-hack unchanged).
- `cargo nextest run -p djinn-db -p djinn-provider --all-features`:
  2278 passed, 0 failed.
