+++ server/crates/djinn-db/src/repositories/mod.rs
@@ -7,6 +7,8 @@ pub mod chat_interruption_notice;
 pub mod code_chunk;
 pub mod commit_file_changes;
 pub mod coordinator_incarnation;
+pub mod credential;
+pub mod custom_provider;
 pub mod dispatch_pause;
 pub mod dispatch_state;
 pub mod doctor_finding;
