+++ server/crates/djinn-provider/Cargo.toml
@@ -35,7 +35,6 @@ serde_json = "1"
 serde_yaml = "0.9"
 serde_urlencoded = "0.7"
 sha2 = "0.10"
-sqlx = { version = "0.8", features = ["mysql", "runtime-tokio-rustls", "macros"] }
 thiserror = "2"
 tokenizers = "0.22.2"
 tokio = { version = "1", features = ["full"] }
